feat: centralize factories
