+++ backend/api/tests/views/test_views_computetask.py
@@ -771,43 +771,6 @@ class GenericTaskViewTests(ComputeTaskViewTests):
         assert data["count"] == 1
 
 
-@pytest.fixture
-def create_compute_task():
-    def _create_compute_task(compute_plan, n_data_sample=4):
-        data_manager = factory.create_datamanager()
-        data_samples = [factory.create_datasample([data_manager]) for _ in range(n_data_sample)]
-        input_keys = {
-            "opener": [data_manager.key],
-            "datasamples": [data_sample.key for data_sample in data_samples],
-        }
-        function = factory.create_function(
-            inputs=factory.build_function_inputs(["datasamples", "opener", "model"]),
-            outputs=factory.build_function_outputs(["model"]),
-            name="simple function",
-        )
-        return factory.create_computetask(
-            compute_plan,
-            function,
-            inputs=factory.build_computetask_inputs(function, input_keys),
-            outputs=factory.build_computetask_outputs(function),
-            data_manager=data_manager,
-            data_samples=[data_sample.key for data_sample in data_samples],
-            status=ComputeTask.Status.STATUS_DONE,
-        )
-
-    return _create_compute_task
-
-
-@pytest.fixture
-def create_compute_plan(create_compute_task):
-    def _create_compute_plan(n_task=20, n_data_sample=4):
-        compute_plan = factory.create_computeplan()
-        [create_compute_task(compute_plan, n_data_sample=n_data_sample) for _ in range(n_task)]
-        return compute_plan
-
-    return _create_compute_plan
-
-
 @pytest.mark.django_db
 def test_n_plus_one_queries_compute_task_in_compute_plan(authenticated_client, create_compute_plan):
     """
